Modified 518.cpp

+++ 518.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int change(int amount, vector<int>& coins) {
-        std::vector<uint64_t> dp(amount + 1, 0);
+        std::vector<unsigned int> dp(amount + 1, 0);
         dp[0] = 1;
 
         for (const auto& coin : coins) {
